split package and module hierarchy
