+++ effectful-co-log/effectful-co-log.cabal
@@ -32,11 +32,14 @@ common common
   build-depends:
     , base >= 4.13 && <= 4.17
     , effectful-core
+    , effectful
+    , effectful-co-log-core
     , co-log ^>= 0.4.0.1
     , co-log-core
     , containers
     , text
     , bytestring
+    , chronos
   
   if impl(ghc < 8.10)
     build-depends:
@@ -47,6 +50,10 @@ library
   hs-source-dirs:   src
   exposed-modules:  
     , Effectful.Colog
+    , Effectful.Colog.Actions
+    , Effectful.Colog.Concurrent
+    , Effectful.Colog.Message
+    , Effectful.Colog.Pure
 
 test-suite effectful-co-log-test
   import: common
@@ -57,4 +64,3 @@ test-suite effectful-co-log-test
     , effectful-co-log
     , hspec ^>= 2.9.4
     , QuickCheck ^>= 2.14.2
-    , silently ^>=1.2.5.2
